requirements.txt
Django==4.1.2
psycopg2-binary
django-cookiebanner==0.2.7
django-bootstrap5==22.1
django-bootstrap-datepicker-plus==5.0.2
whitenoise
django-allauth